+++ src/components/Availability/subcomponents/DayAvailabilityInputBanner.tsx
@@ -1,45 +1,8 @@
-import { useEffect, useState } from 'react'
 import { Checkbox } from '@mui/material'
 import { TimeSlotInput } from './TimeslotInput'
-import { defaultAvailabilityForm } from '../utils/data'
-import { useSelector } from 'react-redux'
-import { useAppSelector } from 'utils/redux/hooks'
-import { useDispatch } from 'react-redux'
-import {
-  getUserAvailability,
-  setUserAvailability,
-} from 'utils/redux/slices/userSlice'
-import { consolidateAvailability } from './SelectTimeInput'
+import { handleCheck } from '../utils/helpers'
 
 export const DayAvailabilityInputBanner = ({ day, days, setDays }) => {
-  //   make a sep. namespace for avail?
-  const dispatch = useDispatch()
-
-  const handleCheck = e => {
-    // default should NOT be 9-5
-    // bc using copy paste time, then auto populates 9-5 as well as the copied time
-    // default should be an empty array
-    // if check to available, and array is empty, add 9-5
-    // if array is not empty, maintain array state
-    const available = !days[e.target.name].available
-    const newAvailability =
-      days[e.target.name].availability.length > 0
-        ? days[e.target.name].availability
-        : [['9:00 AM', '5:00 PM']]
-    const availability = [...days[e.target.name].availability]
-    // this is redundant?
-    const dayToChange = day
-
-    setDays({
-      ...days,
-      [dayToChange]: {
-        available,
-        availability: newAvailability,
-      },
-    })
-    // dispatch(setUserAvailability(days))
-  }
-
   return (
     <div>
       <div className='day-availability-input-banner'>
@@ -47,7 +10,7 @@ export const DayAvailabilityInputBanner = ({ day, days, setDays }) => {
           <div className='check-day'>
             <Checkbox
               name={day}
-              onChange={handleCheck}
+              onChange={e => handleCheck(e, day, days, setDays)}
               sx={{ color: '#022888', '&.Mui-checked': { color: '#022888' } }}
               checked={days[day].available}
             />
